ecommerce/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-js-asset==2.1.0
djangorestframework==3.14.0
dotenv-python==0.0.1
drf-spectacular==0.26.4
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==19.3.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
Pillow==10.0.0
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
